day8: use 8-bit type for digits
```
part 2                  time:   [65.591 us 65.653 us 65.721 us]
                        change: [-15.875% -14.000% -12.063%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
```

There are only 7 segments.

+++ day8/src/main.rs
@@ -28,7 +28,7 @@ fn part1(input: &str) -> usize {
     count
 }
 
-fn convert(number: &str) -> usize {
+fn convert(number: &str) -> u8 {
     number
         .as_bytes()
         .iter()
@@ -37,9 +37,9 @@ fn convert(number: &str) -> usize {
 
 pub fn part2(input: &str) -> usize {
     let mut sum = 0_usize;
-    let mut numbers = [0_usize; 10];
-    let mut five = [0_usize; 3];
-    let mut six = [0_usize; 3];
+    let mut numbers = [0_u8; 10];
+    let mut five = [0_u8; 3];
+    let mut six = [0_u8; 3];
 
     for line in input.lines() {
         let (patterns, out) = line.split_once('|').unwrap();
